# storage/api-docs/api-docs.json
{"swagger":"2.0","info":{"title":"Gallery Rest API","description":"This is example Gallery REST API for Devops\/Sysadmin test in VivifyIdeas.","contact":{"email":"dev@vivifyideas.com"},"license":{"name":"Apache 2.0","url":"http:\/\/www.apache.org\/licenses\/LICENSE-2.0.html"},"version":"1.0.0"},"basePath":"\/api","schemes":["http","https"],"paths":{"\/login":{"post":{"tags":["Authentication"],"summary":"Login new user","operationId":"loginUser","produces":["application\/json"],"parameters":[{"name":"body","in":"body","schema":{"properties":{"email":{"description":"user email","type":"string"},"password":{"description":"user password","type":"string"}}}}],"responses":{"200":{"description":"User logged in"},"401":{"description":"Not authorized"},"403":{"description":"Forbidden"},"422":{"description":"Validation failed"},"500":{"description":"Internal server error"}}}},"\/register":{"post":{"tags":["Authentication"],"summary":"Register new user","operationId":"createUser","produces":["application\/json"],"parameters":[{"name":"body","in":"body","schema":{"properties":{"first_name":{"description":"user first name","type":"string"},"last_name":{"description":"user last name","type":"string"},"email":{"description":"user email","type":"string"},"password":{"description":"user password","type":"string"},"password_confirmation":{"description":"user password confirm","type":"string"},"terms_and_conditions":{"description":"accepted\/declined terms and conditions","type":"boolean"}}}}],"responses":{"200":{"description":"User registrated"},"401":{"description":"Not authorized"},"403":{"description":"Forbidden"},"422":{"description":"Validation failed"},"500":{"description":"Internal server error"}}}},"\/authors-galleries\/{id}\/{page}\/{term}":{"get":{"tags":["Author Gallery"],"summary":"Get author galleries","operationId":"allAuthorGalleries","produces":["application\/json"],"parameters":[{"name":"page","in":"path","required":true,"type":"string"},{"name":"term","in":"path","required":false,"type":"string"},{"name":"id","in":"path","description":"gallery id","required":true,"type":"integer"}],"responses":{"200":{"description":"Successful operation"},"401":{"description":"Not authorized"},"403":{"description":"Forbidden"},"422":{"description":"Validation failed"},"500":{"description":"Internal server error"}},"security":[{"authorization_token":[]}]}},"\/galleries\/{id}\/comments":{"post":{"tags":["Comment"],"summary":"Create a new comment in gallery","operationId":"createComment","produces":["application\/json"],"parameters":[{"name":"body","in":"body","schema":{"properties":{"body":{"description":"Comment body","type":"string"}}}},{"name":"id","in":"path","description":"gallery id","required":true,"type":"integer"}],"responses":{"201":{"description":"Resource created"},"401":{"description":"Not authorized"},"403":{"description":"Forbidden"},"422":{"description":"Validation failed"},"500":{"description":"Internal server error"}},"security":[{"authorization_token":[]}]}},"\/comments\/{id}":{"delete":{"tags":["Comment"],"summary":"Delete a comment in gallery","operationId":"deleteComment","produces":["application\/json"],"parameters":[{"name":"id","in":"path","description":"comment id","required":true,"type":"integer"}],"responses":{"200":{"description":"Successful operation"},"401":{"description":"Not authorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"422":{"description":"Validation failed"},"500":{"description":"Internal server error"}},"security":[{"authorization_token":[]}]}},"\/all-galleries\/{page}\/{term}":{"get":{"tags":["Gallery"],"summary":"Get all galleries","operationId":"allGalleries","produces":["application\/json"],"parameters":[{"name":"page","in":"path","required":true,"type":"string"},{"name":"term","in":"path","required":false,"type":"string"}],"responses":{"200":{"description":"Successful operation"},"401":{"description":"Not authorized"},"403":{"description":"Forbidden"},"422":{"description":"Validation failed"},"500":{"description":"Internal server error"}},"security":[{"authorization_token":[]}]}},"\/galleries":{"post":{"tags":["Gallery"],"summary":"Create a new gallery","operationId":"createGallery","produces":["application\/json"],"parameters":[{"name":"body","in":"body","schema":{"properties":{"name":{"description":"Gallery name","type":"string"},"description":{"description":"Gallery description","type":"string"},"images":{"description":"Array of image URLs","type":"array","items":{"type":"string"}}}}}],"responses":{"201":{"description":"Resource created"},"401":{"description":"Not authorized"},"403":{"description":"Forbidden"},"422":{"description":"Validation failed"},"500":{"description":"Internal server error"}},"security":[{"authorization_token":[]}]}},"\/galleries\/{id}":{"get":{"tags":["Gallery"],"summary":"Get a single gallery","operationId":"singleGallery","produces":["application\/json"],"parameters":[{"name":"id","in":"path","description":"gallery id","required":true,"type":"integer"}],"responses":{"200":{"description":"Successful operation"},"401":{"description":"Not authorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"422":{"description":"Validation failed"},"500":{"description":"Internal server error"}},"security":[{"authorization_token":[]}]},"put":{"tags":["Gallery"],"summary":"Update a new gallery","operationId":"updateGallery","produces":["application\/json"],"parameters":[{"name":"body","in":"body","schema":{"properties":{"name":{"description":"Gallery name","type":"string"},"description":{"description":"Gallery description","type":"string"},"images":{"description":"Array of image URLs","type":"array","items":{"type":"string"}}}}},{"name":"id","in":"path","description":"gallery id","required":true,"type":"integer"}],"responses":{"201":{"description":"Resource created"},"401":{"description":"Not authorized"},"403":{"description":"Forbidden"},"422":{"description":"Validation failed"},"500":{"description":"Internal server error"}},"security":[{"authorization_token":[]}]},"delete":{"tags":["Gallery"],"summary":"Delete a gallery","operationId":"deleteGallery","produces":["application\/json"],"parameters":[{"name":"id","in":"path","description":"gallery id","required":true,"type":"integer"}],"responses":{"200":{"description":"Successful operation"},"401":{"description":"Not authorized"},"403":{"description":"Forbidden"},"404":{"description":"Not found"},"422":{"description":"Validation failed"},"500":{"description":"Internal server error"}},"security":[{"authorization_token":[]}]}},"\/my-galleries\/{page}\/{term}":{"get":{"tags":["Gallery"],"summary":"Get my galleries","operationId":"galleries","produces":["application\/json"],"parameters":[{"name":"page","in":"path","required":true,"type":"string"},{"name":"term","in":"path","required":false,"type":"string"}],"responses":{"200":{"description":"Successful operation"},"401":{"description":"Not authorized"},"403":{"description":"Forbidden"},"422":{"description":"Validation failed"},"500":{"description":"Internal server error"}},"security":[{"authorization_token":[]}]}}},"definitions":{},"externalDocs":{"description":"Find out more about Swagger","url":"http:\/\/swagger.io"},"securityDefinitions":{"authorization_token":{"type":"apiKey","description":"Value: Bearer {token}","name":"Authorization","in":"header"}}}